Refactor: [id]의 UI와 로직 분리
